requirements.txt
streamlit
ollama
Pillow
airllm
torch
transformers
accelerate
scipy
